--- src/App.js
@@ -1,199 +1,0 @@
-import * as React from "react";
-import "./App.css";
-
-const useSemiPersistentState = (key, initialState) => {
-  const [value, setValue] = React.useState(
-    localStorage.getItem(key) || initialState
-  );
-
-  React.useEffect(() => {
-    localStorage.setItem(key, value);
-  }, [value]);
-
-  return [value, setValue];
-};
-
-const initialStories = [
-  {
-    title: "React",
-    url: "https://reactjs.org/",
-    author: "Jordan Walke",
-    num_comments: 3,
-    points: 4,
-    objectID: 0,
-  },
-  {
-    title: "Redux",
-    url: "https://redux.js.org/",
-    author: "Dan Abramov, Andrew Clark",
-    num_comments: 2,
-    points: 5,
-    objectID: 1,
-  },
-];
-
-const getAsyncStories = () =>
-  new Promise((resolve) =>
-    setTimeout(() => resolve({ data: { stories: initialStories } }), 2000)
-  );
-
-const storiesReducer = (state, action) => {
-  switch (action.type) {
-    case "SET_STORIES":
-      return action.payload;
-    case "REMOVE_STORY":
-      return state.filter(
-        (story) => action.payload.objectID !== story.objectID
-      );
-    default:
-      throw new Error();
-  }
-};
-
-const App = () => {
-  const [searchTerm, setSearchTerm] = useSemiPersistentState("search", "");
-  const [stories, dispatchStories] = React.useReducer(storiesReducer, []);
-  const [isLoading, setIsLoading] = React.useState(false);
-  const [isError, setIsError] = React.useState(false);
-
-  React.useEffect(() => {
-    setIsLoading(true);
-
-    getAsyncStories()
-      .then((result) => {
-        dispatchStories({
-          type: "SET_STORIES",
-          payload: result.data.stories,
-        });
-        setIsLoading(false);
-      })
-      .catch(() => setIsError(true));
-  }, []);
-
-  console.log("App renders");
-  //A
-  const handleSearch = (event) => {
-    //C
-    setSearchTerm(event.target.value);
-  };
-
-  const handleRemoveStory = (item) => {
-    console.log(item);
-    dispatchStories({
-      type: "REMOVE_STORY",
-      payload: item,
-    });
-  };
-
-  const searchedStories = stories.filter((story) =>
-    story.title.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  return (
-    <>
-      <h1>My Hacker Stories</h1>
-
-      {/* // B */}
-      <InputWithLabel
-        id="search"
-        value={searchTerm}
-        isFocused
-        onInputChange={handleSearch}
-      >
-        <strong>Search:</strong>
-      </InputWithLabel>
-      <p>
-        Searching for <strong>{searchTerm}</strong>
-      </p>
-
-      <hr />
-      {isError && <p>Something went wrong...</p>}
-
-      {isLoading ? (
-        <p>Loading...</p>
-      ) : (
-        <List list={searchedStories} onRemoveItem={handleRemoveStory} />
-      )}
-    </>
-  );
-};
-
-const InputWithLabel = ({
-  id,
-  value,
-  type = "text",
-  onInputChange,
-  isFocused,
-  children,
-}) => {
-  //A
-  const inputRef = React.useRef();
-  console.log(inputRef);
-  console.log(inputRef.current);
-  console.log(isFocused);
-  //C
-  React.useEffect(() => {
-    if (isFocused && inputRef.current) {
-      //D
-      inputRef.current.focus();
-      console.log(isFocused);
-      console.log("test");
-    }
-    console.log(isFocused);
-    console.log("test2");
-  }, [isFocused]);
-  console.log("Search renders");
-
-  return (
-    <>
-      <label htmlFor={id}>{children} </label>
-      &nbsp;
-      {/* B */}
-      <input
-        ref={inputRef}
-        id={id}
-        type={type}
-        value={value}
-        onChange={onInputChange}
-      />
-    </>
-  );
-};
-
-const List = ({ list, onRemoveItem }) => {
-  console.log("List renders");
-  return (
-    <ul>
-      {list.map((item) => (
-        <Item key={item.objectID} item={item} onRemoveItem={onRemoveItem} />
-      ))}
-    </ul>
-  );
-};
-
-const Item = ({ item, onRemoveItem }) => {
-  /* {url, author, num_comments, points} = props.item */
-
-  return (
-    <li>
-      <span>
-        <a href={item.url}>{item.title}</a>
-      </span>
-      <span>{item.author}</span>
-      <span>{item.num_comments}</span>
-      <span>{item.points}</span>
-      <span>
-        <button
-          type="button"
-          onClick={() => {
-            onRemoveItem(item);
-          }}
-        >
-          Dismiss
-        </button>
-      </span>
-    </li>
-  );
-};
-
-export default App;
