+++ src/Components/InputWithLabel.js
@@ -1,0 +1,40 @@
+import React from "react";
+
+const InputWithLabel = ({
+  id,
+  value,
+  type = "text",
+  onInputChange,
+  isFocused,
+  children,
+}) => {
+  //A
+  const inputRef = React.useRef();
+
+  //C
+  React.useEffect(() => {
+    if (isFocused && inputRef.current) {
+      //D
+      inputRef.current.focus();
+    }
+  }, [isFocused]);
+
+  console.log("Search renders");
+
+  return (
+    <>
+      <label htmlFor={id}>{children} </label>
+      &nbsp;
+      {/* B */}
+      <input
+        ref={inputRef}
+        id={id}
+        type={type}
+        value={value}
+        onChange={onInputChange}
+      />
+    </>
+  );
+};
+
+export default InputWithLabel;
