+++ src/Components/Item.js
@@ -1,0 +1,28 @@
+import React from "react";
+
+const Item = ({ item, onRemoveItem }) => {
+  /* {url, author, num_comments, points} = props.item */
+
+  return (
+    <li>
+      <span>
+        <a href={item.url}>{item.title}</a>
+      </span>
+      <span>{item.author}</span>
+      <span>{item.num_comments}</span>
+      <span>{item.points}</span>
+      <span>
+        <button
+          type="button"
+          onClick={() => {
+            onRemoveItem(item);
+          }}
+        >
+          Dismiss
+        </button>
+      </span>
+    </li>
+  );
+};
+
+export default Item;
