+++ src/Components/List.js
@@ -1,0 +1,19 @@
+import React from "react";
+import Item from "./Item";
+
+const List = ({ list, onRemoveItem }) => {
+  console.log("List renders");
+  return (
+    <ul>
+      {list.map((item) => (
+        <Item 
+          key={item.objectID} 
+          item={item} 
+          onRemoveItem={onRemoveItem} 
+        />
+      ))}
+    </ul>
+  );
+};
+
+export default List;
